Refactor once again
